Filters should generally not indicate default values in the summary, but there are some cases where it can be useful
An example where we'd want to show the default is a list view that only shows a subset by default. For example, only show active products or only current/forthcoming events. If you want users to be aware they are only seeing a subset, you can force visibility of the default value by specifying `exclude_default_value_from_summary=False` on a filter.